Split tokenizers into their own file
