+++ keras_nlp/models/bert/bert_presets_test.py
@@ -20,7 +20,7 @@ from absl.testing import parameterized
 from keras_nlp.models.bert.bert_backbone import BertBackbone
 from keras_nlp.models.bert.bert_classifier import BertClassifier
 from keras_nlp.models.bert.bert_preprocessor import BertPreprocessor
-from keras_nlp.models.bert.bert_preprocessor import BertTokenizer
+from keras_nlp.models.bert.bert_tokenizer import BertTokenizer
 
 
 @pytest.mark.large
